Add joint distribution
